feat: implement voting functionality for stories
This commit introduces the voting functionality for stories, allowing users to upvote or downvote stories. It includes the following changes:

- Added `upvote_count`, `downvote_count`, `vote_count`, and `vote_score` fields to the `stories` table to track vote statistics.
- Created a `voters` relationship on the `Story` and `User` models to manage voters and voted stories.
- Implemented the `currentUserVote` method on the `Story` model to determine the current user's vote on a story.
- Added scopes to `Story` model for ordering stories by vote score, upvotes, and downvotes.
- Created the `vote` method on the `Story` model to allow users to vote on a story, change their vote, or remove their vote.
- Implemented automatic recalculation of vote counts and score when votes are created, updated, or deleted. A fractional penalty is applied to downvotes when calculating the score.
- Implemented unit tests for the new features and methods.

This functionality allows users to express their opinions on stories, improving the overall user experience and enabling better content discovery.

+++ database/migrations/2025_04_22_120528_add_vote_to_stories_table.php
@@ -12,10 +12,23 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('stories', function (Blueprint $table) {
-            $table->unsignedBigInteger('upvote_count')->default(0)->after('view_count')->index();
-            $table->unsignedBigInteger('downvote_count')->default(0)->after('upvote_count')->index();
-            $table->unsignedBigInteger('vote_count')->default(0)->after('downvote_count')->index();
-            $table->unsignedBigInteger('vote_score')->default(0)->after('vote_count')->index();
+            $table->unsignedBigInteger('upvote_count')
+                ->default(0)
+                ->after('view_count')
+                ->index();
+            $table->unsignedBigInteger('downvote_count')
+                ->default(0)
+                ->after('upvote_count')
+                ->index();
+            $table->unsignedBigInteger('vote_count')
+                ->default(0)
+                ->after('downvote_count')
+                ->index();
+            $table
+                ->decimal('vote_score', 8, 2)
+                ->default(0)
+                ->after('vote_count')
+                ->index();
         });
     }
